refactor(infer): hoist brace-repair seam into speculative.rs as ResolutionDoc
Behavior-neutral move: ResolutionDoc (nee LambdaResolutionDoc), the lambda
tree gate, and repaired_doc_at now live beside the marker-insertion parse —
speculative.rs hosts both transient-healed-doc constructors. Seam:
speculative::lambda_doc_at.
